# api/OpenApiSpec.yml
openapi: "3.0.0"
info:
  title: "htv-api"
  version: "1.0.0"
paths:
  /_authcheck:
    get:
      description: "Check if access token is valid"
      tags:
        - "Auth"
      responses:
        200:
          description: "Access token is valid"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  auth:
                    type: "boolean"
                  id:
                    type: "string"
        401:
          description: "Unauthorized"
  /auth/login:
    post:
      description: "Get access token"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
                - "password"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "User's email address"
                password:
                  type: "string"
                  description: "User's password"
                token:
                  type: "string"
                  minLength: 6
                  maxLength: 6
                  description: "Optional MFA code if enabled for account"
      responses:
        200:
          description: "Successfully logged in"
          content:
            application/json:
              schema:
                type: "object"
                $ref: "#/components/schemas/AccessToken"
        401:
          description: "Invalid credentials or MFA code"
        403:
          description: "MFA code required but not provided"
        422:
          description: "Validation errors"
  /auth/sign-up:
    post:
      description: "Register a new user account"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
                - "password"
                - "firstName"
                - "tos"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "User's email address"
                password:
                  type: "string"
                  description: "User's password (must meet strength requirements)"
                firstName:
                  type: "string"
                  description: "User's first name"
                lastName:
                  type: "string"
                  description: "User's last name (optional)"
                groupName:
                  type: "string"
                  description: "Name for the user's default group (optional)"
                tos:
                  type: "boolean"
                  description: "Acceptance of Terms of Service"
      responses:
        200:
          description: "Successfully registered"
          content:
            application/json:
              schema:
                type: "object"
                $ref: "#/components/schemas/AccessToken"
        422:
          description: "Validation errors (email taken, weak password, etc.)"
  /auth/verify-email/{emailVerificationKey}:
    get:
      description: "Verify a user's email address using the verification key"
      tags:
        - "Auth"
      parameters:
        - in: "path"
          name: "emailVerificationKey"
          required: true
          schema:
            type: "string"
          description: "Email verification key sent to the user's email"
        - in: "query"
          name: "redirect"
          schema:
            type: "string"
            enum:
              - "1"
          description: "If set to '1', redirects to frontend after verification"
      responses:
        200:
          description: "Email successfully verified"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  verified:
                    type: "boolean"
                  id:
                    type: "string"
  /auth/forgot:
    post:
      description: "Request a password reset link"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "User's email address"
      responses:
        200:
          description: "Password reset email sent successfully"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "boolean"
  /auth/get-user-by-reset-key/{passwordResetKey}:
    get:
      description: "Get user information by password reset key"
      tags:
        - "Auth"
      parameters:
        - in: "path"
          name: "passwordResetKey"
          required: true
          schema:
            type: "string"
          description: "Password reset key sent to the user's email"
      responses:
        200:
          description: "Returns user information"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "string"
                  email:
                    type: "string"
                    format: "email"
  /auth/reset:
    post:
      description: "Reset user password using reset key"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
                - "password"
                - "passwordResetKey"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "User's email address"
                password:
                  type: "string"
                  description: "New password (must meet strength requirements)"
                passwordResetKey:
                  type: "string"
                  description: "Password reset key sent to the user's email"
      responses:
        200:
          description: "Password successfully reset and user logged in"
          content:
            application/json:
              schema:
                type: "object"
                $ref: "#/components/schemas/AccessToken"
  /auth/get-user-by-invite-key/{inviteKey}:
    get:
      description: "Get user information by invite key"
      tags:
        - "Auth"
      parameters:
        - in: "path"
          name: "inviteKey"
          required: true
          schema:
            type: "string"
          description: "Invitation key sent to the user's email"
      responses:
        200:
          description: "Returns basic user information"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "string"
                    description: "User ID"
                  email:
                    type: "string"
                    format: "email"
                    description: "User's email address"
  /auth/invite:
    post:
      description: "Accept an invitation and create a user account"
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "email"
                - "password"
                - "firstName"
                - "tos"
                - "inviteKey"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "User's email address"
                password:
                  type: "string"
                  description: "User's password (must meet strength requirements)"
                firstName:
                  type: "string"
                  description: "User's first name"
                lastName:
                  type: "string"
                  description: "User's last name (optional)"
                tos:
                  type: "boolean"
                  description: "Acceptance of Terms of Service"
                inviteKey:
                  type: "string"
                  description: "Invitation key received via email"
      responses:
        200:
          description: "Successfully accepted invitation and created account"
          content:
            application/json:
              schema:
                type: "object"
                $ref: "#/components/schemas/AccessToken"
...
components:
  schemas:
...
    AccessToken:
      properties:
        accessToken:
          type: "string"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...